owndrive/requirements.txt
Django==1.9.2
djangorestframework==3.3.2
mod-wsgi==4.4.22
MySQL-python==1.2.5
